Renamed model to "SchemeModel"
